Add Entity, Repository, Service, ServiceImpl, Controller and apply validation on DTO

+++ src/main/java/com/blogapplication/repository/CommentRepo.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.blogapplication.entity.Comment;
 
-public interface CommentRepo  extends JpaRepository<Comment, Integer> {
+public interface CommentRepo  extends JpaRepository<Comment, Long> {
